+++ programs/rpg/src/constants.rs
@@ -1,0 +1,9 @@
+pub const IS_FROZEN_FLAG: u8 = 1 << 0;
+pub const IS_POISONED_FLAG: u8 = 1 << 1;
+pub const IS_BURNING_FLAG: u8 = 1 << 2;
+pub const IS_BLESSED_FLAG: u8 = 1 << 3;
+pub const IS_CURSED_FLAG: u8 = 1 << 4;
+pub const IS_STUNNED_FLAG: u8 = 1 << 5;
+pub const IS_SLOWED_FLAG: u8 = 1 << 6;
+pub const IS_BLEEDING_FLAG: u8 = 1 << 7;
+pub const NO_EFFECT_FLAG: u8 = 0b00000000;
